feat(all): new template registry and global loader
This is a BREAKING CHANGE. The new base loader manages a template
registry. This is uses to properly track loaded templates, their
dependencies, the final view resources associated with each template
and the factory that instantiates views from the template. This is a
foundational pieces used by concrete loaders. The default-loader
leverages this as part of the way it plugs into the system and
require.js loader.

This commit also the old createDefaultLoader api in favor of using a
global variable for this so that it’s easier to re-use the bootstrapper.

Finally, we have changes the name of the “import” element to
“au-import” This makes it clearer that this is part of Aurelia as well
enables supporting IE9.