perf(loader): replace all querySelectorAll methods with getElementsByTagName
The loader was looking up elements in the DOM, more specifically <template> and <require> using document.querySelectorAll. While this method works, the querySelectorAll method returns a StaticNodeList which means the matched elements are matched upfront. This is in comparison to getElementsByTagName which returns a DynamicNodeList which is a living representation of the DOM and thus faster as no pre-computation is performed.